Enable sending caps over IPC
This refactors the serial driver to use this mechanism - instead of
having a "read" message that returns data and a signal notification that
interested threads can wait on separately, we now have interested
readers register an endpoint for the serial driver to send data over.
I do need to achnowlege that since the serial driver would be blocked if
no one was recv'ing its message, so I'll intend to add non-blocking
send and recv in the near future and potentially investigate extending
L4-style IPC to support a "select" mechanism to allow a thread to wait
for multiple different kinds of event.