Clean up message_info_t and reg params on endpoint
endpoint and notification had a bad habit of reaching into the receiver
thread's saved stack frame and that had been causing problems for as
long as this system has existed. This change refactors the resulting tag
and badge to always be passed through the ipc_buffer and then not return
message_info_t back to their callers.