Add pci manager program
This program is designed to provide restricted permission semantics to
PCI accesses, holders of endpoint capabilities to this program with
badge 0 can issue capabilities restricted to PCI read and write
operations on specific PCI addresses.

This isn't currently spawned or finished, but it does compile and is
sufficient to commit.