Claim 2 more bits from cap pointers
Assert that capability pointers are always 16-byte aligned; previously
we only claimed the bottom 2 bits for capability data. The smallest
kernel object pointed to by a capability pointer anywhere in the kernel
is 16 bytes, so we can use two more bits for us.

This commit also claims the original lowest-order reserved bit for an
is_original bit. I'm putting this here as opposed to on the derivation
tree because it does affect capability behavior (delete on an original
capability destroys the kernel object) so I think it's semantically part
of the capability itself whether it's original or not.